Refactor SaToken config to memory mode with Redis support
