--- shanhaifangzhou-admin/src/main/java/com/shanhaifangzhou/admin/config/SimpleSaTokenConfig.java
@@ -1,46 +1,0 @@
-package com.shanhaifangzhou.admin.config;
-
-import cn.dev33.satoken.config.SaTokenConfig;
-import cn.dev33.satoken.dao.SaTokenDao;
-import cn.dev33.satoken.dao.SaTokenDaoDefaultImpl;
-import cn.dev33.satoken.interceptor.SaInterceptor;
-import cn.dev33.satoken.router.SaRouter;
-import cn.dev33.satoken.stp.StpUtil;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-/**
- * 简化Sa-Token配置 - 使用内存模式
- *
- * @author shanhaifangzhou
- * @since 2024-01-01
- */
-@Configuration
-public class SimpleSaTokenConfig implements WebMvcConfigurer {
-
-    /**
-     * 配置Sa-Token使用内存模式
-     */
-    @Bean
-    @Primary
-    public SaTokenDao saTokenDao() {
-        return new SaTokenDaoDefaultImpl();
-    }
-
-    /**
-     * 注册Sa-Token拦截器
-     */
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        // 注册Sa-Token拦截器，打开注解式鉴权功能
-        registry.addInterceptor(new SaInterceptor(handle -> {
-            // 指定需要登录认证的路径
-            SaRouter.match("/**")
-                    .notMatch("/auth/login", "/auth/captcha", "/test/**", "/system/info", "/doc.html", "/webjars/**", "/v3/api-docs/**")
-                    .check(r -> StpUtil.checkLogin());
-        })).addPathPatterns("/**");
-    }
-}
