+++ shanhaifangzhou-admin/src/main/resources/application-simple.yml
@@ -6,6 +6,18 @@ server:
 spring:
   application:
     name: shanhaifangzhou-admin
+  data:
+    redis:
+      host: localhost
+      port: 6379
+      database: 0
+      timeout: 3000ms
+      lettuce:
+        pool:
+          max-active: 8
+          max-wait: -1ms
+          max-idle: 8
+          min-idle: 0
 
 # Sa-Token配置 - 使用内存模式
 sa-token:
